requirements.txt
openai==0.28.1
langchain==0.2.0
python-dotenv==1.0.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-ollama==0.3.6
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
ollama==0.5.1
